Continuum o Upgrading to JDO 2. Going with Modello alpha-3-SNAPSHOT.
Continuum Model
o Making the configuration a part of the project. These are really
  deprecated and should go away.
o The ContinuumBuildResult is a part of the ContinuumBuild.
o The UpdateScmResult is a part of the ContinuumBuild.
o The ScmFile collections are a part of the CheckedOutScmResult and
  UpdateScmResult.
o The ContinuumProject object and other objects that has subclasses should
  be made abstract.

Continuum API
o Changed addMavenOneProject( url ) and addMavenTwoProject( url ) to
  return the ContinuumProjectBuildingResult object. This was done to give
  the client access to the result of the adding operation. Currently it
  only uses it to get hang of the project ids added. Changed
  addXXXProject( XXXProject ) to return the project id of the project
  added.

  All of these methods should return the same object encapsulating the
  result of the operation but it should probably not be
  ContinuumProjectBuildingResult objects. TBD.

Continuum Core
o Depracating methods that not work directly on a model class. These
  should be removed before the alpha. Removed any references to them where
  this was easy to do. Still need a pass to clean it up completly thus
  they're still there.

o Removed lots of JDO test configurations. Changed the configuration in
  src/main/resources to contain a configuration useful for testing. This
  should probably be moved to src/test/resources and the core should by
  default miss this configuration.

Continuum XML-RPC Interface
o Removing methods only available though direct usage of ContinuumCore in
  the XML-RPC client. The XML-RPC interface now directly relates to the
  Contiuum component only.

o Updating the excludes when transforming a object to a hashtable as JDO
  now complains if we're accessing fields that's not detached instead of
  returning null or just a empty collection.

Continuum XML-RPC Python Integration Tests and Client
o Reworked the continuum module. It now contains a Continuum class that
  can be instanciated with a URL. This increases resusability of this
  module.  Reorganized the methods to match the XML-RPC interface.

o Adding AntProject as this is required when adding Ant projects.
  Currently adding Maven 1 and 2 projects though a URL so those and the
  shell project will be added later.

o Updating the continuum Python module to match the XML-RPC interface one
  to one. Updating the ITs to match the new client API.

  The Ant and Shell parts of the ITs are currenly broken as they're not
  properly implemented in the core.


git-svn-id: https://svn.apache.org/repos/asf/maven/continuum/trunk@190240 13f79535-47bb-0310-9956-ffa450edef68