this is a fix for something discovered in plxredback-66
the security setup on continuum stems from the ContinuumActionSupport
class which is the parent to most actions in continuum.  it is also where
all of the authz goop is abstracted away too, and it uses the prepare
interceptor to put the securitySession of the user into action.  Well
the xwork interceptor stack was setup so that this happened _before_ the 
autologin interceptors were calls, which in effect made the first
page on the autologin path fail the authorizations of that action...

this was masked because once that other interceptor ran it was all setup to
go so jsp's rendering on the page actually had the relevant information 
available.

moral of the story, get your interceptor stack in order...

git-svn-id: https://svn.apache.org/repos/asf/maven/continuum/trunk@519174 13f79535-47bb-0310-9956-ffa450edef68